Fixed module path separator in Windows (#116)
* Fixed module path separator in Windows

Module path separator in Windows is currently "/" ie: dw/util/Math, which
is incorrect.  It should be "::" (ie: dw::util::Math).  This is because
Intellij's virtual file uses "/" separator, but Dataweave parser SDK expects
the native file separator, which is "\" in Windows.

Pre-correcting the file separator before sending the path to the parser SDK
fixes this issue.

This bug caused modules and some functions as being not resolvable in the editor.
Also, the auto suggest gives the wrong path (ie: suggesting dw/util/Math
instead of dw::util::Math).

* Fixed per code review