Fixed plugin crash and add libraries
In the recent commit, there was a full dispose that caused the
WeavePreviewComponent not being able to be reused.  This causes plugin
crash in the IDE when opening the Dataweave preview.  The proposed
solution is to recreate the WeavePreviewComponent when setting a new
 DWL file, because WeavePreviewComponent is always in a disposed state
at this time.

In a dataweave that has java import for classes that live in a library,
the Dataweave preview complains that it cannot find the class file.
The proposed solution is to send all libraries, except SDK, to the
agent so that the class file is available when the script engine
evaluates the Dataweave in the agent.